feat(tooltips): allow enabling and disabling of tooltips
closes #5
In the angular material version 1.0.1 a bug was introduced. When using `md-menu` together with `md-tooltip` unwanted behaviour appeared
(see: https://github.com/angular/material/issues/6662). Until this is fixed we enabled the toggling of tooltips. Per default the tooltips are disabled.
When the problem is solved by angular material you can just set `show-tooltips` to `true` to show tooltips again.